Refactor bulk account fetch
